requirements.txt
alembic==1.14.1
Flask==3.1.0
Flask_Login==0.6.3
Flask_Migrate==4.1.0
Flask_Moment==1.0.6
Flask_SocketIO==5.5.1
flask_sqlalchemy==3.1.1
gevent==24.11.1
openai==1.75.0
python-dotenv==1.1.0
SQLAlchemy==2.0.37
Werkzeug==3.1.3
